msm: kgsl: Allow internal commands from detached context
If a context is active and we try to detch the context then the
commands to switch out the context will return error since we
set the detached bit of the context before submitting the
commands to switch it out. Fix this by allowing the internal
kgsl commands to be submitted even if the context is detached.

Change-Id: I294b9ac151085a53dc989915e3a612a0ebfbacc4
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>