msm: kgsl: Revert the GPU address / ID conflict fix
Partial revert of 387a5dfcf91000155c3e3eec469f1f4871446864. The IDs
are now big enough that 32 bit user space ends up sign-extending
the offset and causing all sorts of confusion in the kernel. This
is a partial revert because the previous change also had a rather
critical fix to avoid OOM conditions and we would rather keep
that around.

Change-Id: Ic0dedbad4a5ac8f35d97da91b92f6d3d2951403d
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>